+++ client/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
+- Removed the Add Row button from single field options in advanced search ([#149](https://github.com/project-lux/lux-frontend/issues/149)).
+
 ### Fixed
 
 - Fixed hierarchy on object pages ([#94](https://github.com/project-lux/lux-frontend/issues/94)).